+++ src/Core/Signer/SignerV4.php
@@ -4,7 +4,10 @@ namespace AsyncAws\Core\Signer;
 
 use AsyncAws\Core\Credentials\Credentials;
 use AsyncAws\Core\Exception\InvalidArgument;
-use AsyncAws\Core\Exception\RuntimeException;
+use AsyncAws\Core\Stream\FixedSizeStream;
+use AsyncAws\Core\Stream\IterableStream;
+use AsyncAws\Core\Stream\Stream;
+use AsyncAws\Core\Stream\StringStream;
 
 /**
  * Version4 of signer.
@@ -54,14 +57,7 @@ class SignerV4 implements Signer
 
     public function sign(Request $request, ?Credentials $credentials): void
     {
-        $body = $request->getBody() ?? '';
-        if (\is_resource($body) && -1 === fseek($body, 0)) {
-            throw new RuntimeException('Unable to seek the resource');
-        }
-
         if (null === $credentials) {
-            $request->setHeader('content-length', $this->getContentLength($request));
-
             return;
         }
 
@@ -86,30 +82,22 @@ class SignerV4 implements Signer
 
         $canonicalHeaders = $this->getCanonicalizedHeaders($request);
 
-        $canonicalRequest = implode(
-            "\n",
-            [
-                $request->getMethod(),
-                $this->getCanonicalizedPath($parsedUrl),
-                $this->getCanonicalizedQuery($parsedUrl),
-                \implode("\n", array_values($canonicalHeaders)),
-                '', // empty line after headers
-                implode(';', \array_keys($canonicalHeaders)),
-                $request->getHeader('x-amz-content-sha256'),
-            ]
-        );
+        $canonicalRequest = implode("\n", [
+            $request->getMethod(),
+            $this->getCanonicalizedPath($parsedUrl),
+            $this->getCanonicalizedQuery($parsedUrl),
+            \implode("\n", array_values($canonicalHeaders)),
+            '', // empty line after headers
+            implode(';', \array_keys($canonicalHeaders)),
+            $request->getHeader('x-amz-content-sha256'),
+        ]);
 
-        $stringToSign = implode(
-            "\n",
-            [
-                self::ALGORITHM_REQUEST,
-                $amzDate,
-                implode('/', $credentialScope),
-                hash('sha256', $canonicalRequest),
-            ]
-        );
-
-        $signature = hash_hmac('sha256', $stringToSign, $signingKey);
+        $signature = hash_hmac('sha256', implode("\n", [
+            self::ALGORITHM_REQUEST,
+            $amzDate,
+            implode('/', $credentialScope),
+            hash('sha256', $canonicalRequest),
+        ]), $signingKey);
 
         $authorizationHeader = sprintf(
             '%s Credential=%s/%s, SignedHeaders=%s, Signature=%s',
@@ -123,122 +111,69 @@ class SignerV4 implements Signer
         $request->setHeader('authorization', $authorizationHeader);
     }
 
-    private function getContentLength(Request $request): ?int
-    {
-        if ($request->hasHeader('content-length')) {
-            return (int) ((array) $request->getHeader('content-length'))[0];
-        }
-
-        $body = $request->getBody();
-        if (\is_string($body)) {
-            return \strlen($body);
-        }
-
-        if (\is_resource($body)) {
-            return fstat($body)['size'] ?? null;
-        }
-
-        return null;
-    }
-
     private function prepareBody(Request $request, string $amzDate, string $credentialScope, string &$signature, string $signingKey): void
     {
         $body = $request->getBody();
 
-        // we can't manage signature of undefined length closure
-        $contentLength = $this->getContentLength($request);
-        if (null === $contentLength) {
-            if (\is_callable($body)) {
-                $buffer = '';
-                while (true) {
-                    if (!\is_string($data = $body(self::CHUNK_SIZE))) {
-                        throw new InvalidArgument(sprintf('The return value of the "body" option callback must be a string, %s returned.', \gettype($data)));
-                    }
-
-                    if ('' == $data) {
-                        break;
-                    }
-
-                    $buffer .= $data;
-                }
-
-                $request->setBody($body = $buffer);
-            } elseif (\is_resource($body)) {
-                $request->setBody($body = \stream_get_contents($body));
-            }
+        if ($request->hasHeader('content-length')) {
+            $contentLength = (int) ((array) $request->getHeader('content-length'))[0];
+        } else {
+            $contentLength = $body->length();
         }
 
-        if (\is_string($body)) {
-            $request->setHeader('x-amz-content-sha256', hash('sha256', $body));
+        // we can't manage signature of undefined length. Let's convert it to string
+        if (null === $contentLength) {
+            $request->setBody($body = StringStream::create($body));
+            $contentLength = $body->length();
+        }
+
+        // no need to stream small body
+        if ($contentLength < self::CHUNK_SIZE) {
+            $request->setBody($body = StringStream::create($body));
+            $request->setHeader('x-amz-content-sha256', hash('sha256', $body->stringify()));
             $request->setHeader('content-length', $contentLength);
 
             return;
-        }
-
-        $streamReader = null;
-        if (\is_callable($body)) {
-            $eof = false;
-            $buffer = '';
-            $streamReader = static function () use ($body, &$buffer, &$eof): string {
-                while (!$eof && \strlen($buffer) < self::CHUNK_SIZE) {
-                    if (!\is_string($data = $body(self::CHUNK_SIZE))) {
-                        throw new InvalidArgument(sprintf('The return value of the "body" option callback must be a string, %s returned.', \gettype($data)));
-                    }
-
-                    $buffer .= $data;
-                    $eof = '' === $data;
-                }
-
-                $data = substr($buffer, 0, self::CHUNK_SIZE);
-                $buffer = substr($buffer, self::CHUNK_SIZE);
-
-                return $data;
-            };
-        } elseif (\is_resource($body)) {
-            $streamReader = static function () use ($body): string {
-                return \fread($body, self::CHUNK_SIZE);
-            };
-        }
-        if (null === $contentLength) {
-            throw new RuntimeException('Unable to get resource size');
-        }
-        if (null === $streamReader) {
-            throw new InvalidArgument(\sprintf('Unexpected body "%s".', \is_object($body) ? \get_class($body) : \gettype($body)));
         }
 
         $request->setHeader('content-encoding', 'aws-chunked');
         $request->setHeader('x-amz-decoded-content-length', $contentLength);
         $request->setHeader('x-amz-content-sha256', 'STREAMING-' . self::ALGORITHM_CHUNK);
 
+        // Compute size of content + metadata used sign each Chunk
         $chunkCount = (int) ceil($contentLength / self::CHUNK_SIZE);
         $fullChunkCount = $chunkCount * self::CHUNK_SIZE === $contentLength ? $chunkCount : ($chunkCount - 1);
-
         $metaLength = \strlen(";chunk-signature=\r\n\r\n") + 64;
-        $contentLength = $contentLength + $fullChunkCount * ($metaLength + \strlen((string) dechex(self::CHUNK_SIZE))) + ($chunkCount - $fullChunkCount) * ($metaLength + \strlen((string) dechex($contentLength % self::CHUNK_SIZE))) + $metaLength + 1;
-        $request->setHeader('content-length', $contentLength);
+        $request->setHeader('content-length', $contentLength + $fullChunkCount * ($metaLength + \strlen((string) dechex(self::CHUNK_SIZE))) + ($chunkCount - $fullChunkCount) * ($metaLength + \strlen((string) dechex($contentLength % self::CHUNK_SIZE))) + $metaLength + 1);
 
-        $last = false;
-        $streamBody = static function () use ($streamReader, $amzDate, $credentialScope, &$signature, &$last, $signingKey): string {
-            if ($last) {
-                return '';
+        $body = IterableStream::create((static function (Stream $body) use ($amzDate, $credentialScope, $signingKey, &$signature): iterable {
+            $emptyHash = hash('sha256', '');
+            foreach (FixedSizeStream::create($body, self::CHUNK_SIZE) as $chunk) {
+                $signature = hash_hmac('sha256', implode("\n", [
+                    self::ALGORITHM_CHUNK,
+                    $amzDate,
+                    $credentialScope,
+                    $signature,
+                    $emptyHash,
+                    hash('sha256', $chunk),
+                ]), $signingKey);
+
+                yield sprintf("%s;chunk-signature=%s\r\n", dechex(\strlen($chunk)), $signature) . "$chunk\r\n";
             }
 
-            if ('' === $data = $streamReader()) {
-                $last = true;
-            }
-
-            $stringToSign = implode("\n", [
+            $signature = hash_hmac('sha256', implode("\n", [
                 self::ALGORITHM_CHUNK,
                 $amzDate,
                 $credentialScope,
                 $signature,
-                hash('sha256', ''),
-                hash('sha256', $data),
-            ]);
+                $emptyHash,
+                $emptyHash,
+            ]), $signingKey);
 
-            return sprintf('%s;chunk-signature=%s' . "\r\n", dechex(\strlen($data)), $signature = hash_hmac('sha256', $stringToSign, $signingKey)) . $data . "\r\n";
-        };
-        $request->setBody($streamBody);
+            yield sprintf("%s;chunk-signature=%s\r\n\r\n", dechex(0), $signature);
+        })($body));
+
+        $request->setBody($body);
     }
 
     private function getCanonicalizedQuery(array $parseUrl): string